Generate and handle modified emoji strategy
This generates a modified version of the new emoji_strategy.json file
reduced down to only characters that were present in the original
strategy file.

The stemoji.ts script is modified to transform the new/modified strategy
so it continues to work similarly to how it did previously without
re-writing the whole thing.