Move normalisation from dict build to sort only
Instead of normalising the dictionary numbers from number bar plus
letter to plain numbers in the built typey-type.json dictionary, we
instead use normalisation only for sorting purposes. This preserves
letter-based number outlines so that it's easier for beginners to find
which key they need to press by looking at a diagram that is labelled
only with letters and number bar and no numbers.