Add emilys-symbols.json to source data
Because emilys-symbols.json already existed in the target dictionaries
directory and it was not mentioned in the sourceDictionariesIndex.json
file, nothing tried to copy it and nothing complained that it was
missing. However, running `make clean` then `build-everything` would hit
an error so adding `emilys-symbols.json` in the correct source locations
where it should have already been fixes that issue.

Update README to add more clarity around source dictionary files and
metadata.